+++ experiment/selinux/src/lib.rs
@@ -1,3 +1,5 @@
-pub mod label;
 pub mod selinux;
+pub mod selinux_label;
 pub mod tools;
+
+pub use selinux::SELinux;
